requirements.txt
deepface==0.0.93
imageio==2.37.0
numpy==2.3.1
opencv_python==4.11.0.86
Pillow==11.3.0
streamlit==1.46.0
tensorflow==2.19.0
uform==3.1.2
